VACMS-6213: Move title setting logic to va_gov_vamc.
