+++ config/sync/core.extension.yml
@@ -179,7 +179,6 @@ module:
   va_gov_post_api: 0
   va_gov_preview: 0
   va_gov_resources_and_support: 0
-  va_gov_root_pages: 0
   va_gov_user: 0
   va_gov_user_guide: 0
   va_gov_vamc: 0
