--- docroot/modules/custom/va_gov_root_pages/va_gov_root_pages.info.yml
@@ -1,5 +1,0 @@
-name: 'VA.gov root pages'
-type: module
-description: 'Contains logic for top level content types.'
-core_version_requirement: ^8 || ^9
-package: 'Custom'
