--- docroot/modules/custom/va_gov_root_pages/va_gov_root_pages.libraries.yml
@@ -1,3 +1,0 @@
-set_menu_title:
-  js:
-    js/set_menu_title.js: {}
