--- docroot/modules/custom/va_gov_root_pages/va_gov_root_pages.module
@@ -1,24 +1,0 @@
-<?php
-
-/**
- * @file
- * Contains va_gov_root_pages.module.
- */
-
-use Drupal\Core\Routing\RouteMatchInterface;
-
-/**
- * Implements hook_help().
- */
-function va_gov_root_pages_help($route_name, RouteMatchInterface $route_match) {
-  switch ($route_name) {
-    // Main module help for the va_gov_root_pages module.
-    case 'help.page.va_gov_root_pages':
-      $output = '';
-      $output .= '<h3>' . t('About') . '</h3>';
-      $output .= '<p>' . t('Contains logic for top level content types.') . '</p>';
-      return $output;
-
-    default:
-  }
-}
