--- docroot/modules/custom/va_gov_root_pages/va_gov_root_pages.services.yml
@@ -1,5 +1,0 @@
-services:
-  va_gov_root_pages.entity_event_subscriber:
-    class: Drupal\va_gov_root_pages\EventSubscriber\EntityEventSubscriber
-    tags:
-      - { name: event_subscriber }
