+++ docroot/modules/custom/va_gov_vamc/css/set_menu_title.css
@@ -1,0 +1,6 @@
+.js-form-item-menu-title.form-item-menu-title #edit-menu-title.disabled-title {
+  background-color: hsla(0, 0%, 0%, 0.08);
+  border-color: #d4d4d4;
+  box-shadow: none;
+  pointer-events: none;
+}
