+++ docroot/modules/custom/va_gov_vamc/va_gov_vamc.libraries.yml
@@ -4,3 +4,9 @@ set_ief_administration_select:
 limit_vamcs_to_workbench:
   js:
     js/limit_vamcs_to_workbench.js: {}
+set_menu_title:
+  css:
+    component:
+      css/set_menu_title.css: {}
+  js:
+    js/set_menu_title.js: {}
